pyproject.toml
[project]
name = "cuisineai"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.2.15",
    "langchain-deepseek>=1.0.1",
    "langchain-openai>=1.2.0",
    "langchain-tavily>=0.2.18",
    "langgraph-checkpoint-sqlite>=3.0.3",
    "python-dotenv>=1.2.2",
]

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple/"
default = true
